backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pillow
numpy
opencv-python-headless
pytesseract
